requirements.txt
fastapi==0.111.0
loguru==0.7.2
pydantic==2.7.1
PyYAML==6.0
redis==4.5.3
Requests==2.32.2
uvicorn==0.29.0
openai==1.30.1
tiktoken~=0.5.2
uuid7
fire
transformers==4.43.2 
diffusers==0.24.0
eventlet==0.35.2
numpy==1.26.4
urllib3==1.26.20
boto3==1.35.63
wget
psutil